+++ matrix-sdk/src/main/java/org/matrix/androidsdk/crypto/keysbackup/KeysBackupPassword.kt
@@ -19,6 +19,7 @@
  */
 package org.matrix.androidsdk.crypto.keysbackup
 
+import android.support.annotation.WorkerThread
 import org.matrix.androidsdk.listeners.ProgressListener
 import org.matrix.androidsdk.util.Log
 import java.util.*
@@ -45,6 +46,7 @@ data class GeneratePrivateKeyResult(
  *
  * @return a {privateKey, salt, iterations} tuple.
  */
+@WorkerThread
 fun generatePrivateKeyWithPassword(password: String, progressListener: ProgressListener?): GeneratePrivateKeyResult {
     val salt = generateSalt()
     val iterations = DEFAULT_ITERATION
@@ -59,17 +61,21 @@ fun generatePrivateKeyWithPassword(password: String, progressListener: ProgressL
  * @param password the password used to generated the private key.
  * @param salt the salt.
  * @param iterations number of key derivations.
+ * @param progressListener the progress listener
  *
  * @return a private key.
  */
+@WorkerThread
 fun retrievePrivateKeyWithPassword(password: String,
                                    salt: String,
-                                   iterations: Int): ByteArray {
-    return deriveKey(password, salt, iterations, null)
+                                   iterations: Int,
+                                   progressListener: ProgressListener? = null): ByteArray {
+    return deriveKey(password, salt, iterations, progressListener)
 }
 
 /**
  * Compute a private key by deriving a password and a salt strings.
+ *
  * @param password the password.
  * @param salt the salt.
  * @param iterations number of derivations.
@@ -77,6 +83,7 @@ fun retrievePrivateKeyWithPassword(password: String,
  *
  * @return a private key.
  */
+@WorkerThread
 private fun deriveKey(password: String,
                       salt: String,
                       iterations: Int,